+++ djangobmf/contrib/product/bmf_module.py
@@ -15,8 +15,8 @@ from .categories import ProductCategory
 from .models import Product
 from .models import ProductTax
 from .models import PRODUCT_SERVICE
-from .views import ProductCreateView
 from .views import ProductDetailView
+from .views import ProductCreateView
 from .views import ProductUpdateView
 
 
@@ -25,8 +25,8 @@ class ProductModule(Module):
     model = Product
     default = True
     create = ProductCreateView
-    detail = ProductDetailView
     update = ProductUpdateView
+    detail = ProductDetailView
 
 
 @register