+++ static_files/js/ui-init.js
@@ -11,6 +11,7 @@ var BMFEVENT_DATA = "bmf.event.update.data";
 var BMFEVENT_MODAL = "bmf.event.update.modal";
 var BMFEVENT_NAVIGATION = "bmf.event.update.navigation";
 var BMFEVENT_OBJECT = "bmf.event.update.object";
+var BMFEVENT_OBJECTDATA = "bmf.event.update.objectdata";
 var BMFEVENT_SIDEBAR = "bmf.event.update.sidebar";
 
 // INIT APP