+++ CHANGELOG.rst
@@ -1,11 +1,12 @@
 Version 0.3.0 (currently in development)
 ========================================
 
-* Related data on detail views
 
 Version 0.2.X
 ========================================
 
+* Related data on detail views
+* Added a CountryField validated by ``pycountry``
 * Added polish translation
 * Added russian translation
 * Rebuild Notifications with angularJS